+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## main branch
 
 - UI: 增加本地缓存进行地址管理
+- worker: 增加 `FORWARD_ADDRESS_LIST` 全局邮件转发地址(等同于 `catch all`)
 
 ## v0.4.6